Update decorator to have proper method for returning a task
